+++ Example/Example/Modules/Demo/Demo.storyboard
@@ -1,0 +1,69 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="14460.31" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="1TA-kC-qwv">
+    <device id="retina4_7" orientation="portrait">
+        <adaptation id="fullscreen"/>
+    </device>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14460.20"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <scenes>
+        <!--Demo View Controller-->
+        <scene sceneID="k6t-ic-mNz">
+            <objects>
+                <viewController id="UJa-Dr-7yO" customClass="DemoViewController" customModule="Example" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="rNy-gw-US6">
+                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="7fF-ak-WWB">
+                                <rect key="frame" x="115" y="318.5" width="145" height="30"/>
+                                <state key="normal" title="Network Reachability"/>
+                                <connections>
+                                    <segue destination="8Qp-Ax-yqr" kind="show" id="Yxx-1C-KaV"/>
+                                </connections>
+                            </button>
+                        </subviews>
+                        <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                        <constraints>
+                            <constraint firstItem="7fF-ak-WWB" firstAttribute="centerY" secondItem="rNy-gw-US6" secondAttribute="centerY" id="0VO-Bo-IgO"/>
+                            <constraint firstItem="7fF-ak-WWB" firstAttribute="centerX" secondItem="rNy-gw-US6" secondAttribute="centerX" id="Rw2-at-xcv"/>
+                        </constraints>
+                        <viewLayoutGuide key="safeArea" id="Pih-wg-Joh"/>
+                    </view>
+                    <navigationItem key="navigationItem" id="YYw-mW-rel"/>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="eGI-nR-xrp" userLabel="First Responder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="912.79999999999995" y="99.400299850074973"/>
+        </scene>
+        <!--NetworkReachability-->
+        <scene sceneID="ePL-kW-2Bd">
+            <objects>
+                <viewControllerPlaceholder storyboardName="NetworkReachability" id="8Qp-Ax-yqr" sceneMemberID="viewController"/>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="lfQ-AN-5S9" userLabel="First Responder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="1570.4000000000001" y="98.950524737631198"/>
+        </scene>
+        <!--Navigation Controller-->
+        <scene sceneID="smg-XL-6Qm">
+            <objects>
+                <navigationController automaticallyAdjustsScrollViewInsets="NO" id="1TA-kC-qwv" sceneMemberID="viewController">
+                    <toolbarItems/>
+                    <navigationBar key="navigationBar" contentMode="scaleToFill" insetsLayoutMarginsFromSafeArea="NO" id="dnF-db-UMj">
+                        <rect key="frame" x="0.0" y="20" width="375" height="44"/>
+                        <autoresizingMask key="autoresizingMask"/>
+                    </navigationBar>
+                    <nil name="viewControllers"/>
+                    <connections>
+                        <segue destination="UJa-Dr-7yO" kind="relationship" relationship="rootViewController" id="g8K-00-L5Y"/>
+                    </connections>
+                </navigationController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="LEe-k7-XkC" userLabel="First Responder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="-26.399999999999999" y="99.400299850074973"/>
+        </scene>
+    </scenes>
+</document>
