+++ Example/Example/Modules/Demo/DemoViewController.swift
@@ -1,0 +1,18 @@
+//
+//  DemoViewController.swift
+//  Example
+//
+//  Created by Dylan Gyesbreghs on 02/01/2019.
+//  Copyright © 2019 icapps. All rights reserved.
+//
+
+import UIKit
+
+class DemoViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        title = "Demo"
+    }
+    
+}
