+++ Example/Example/Supporting Files/Assets.xcassets/AppIcon.appiconset/Contents.json
