+++ Example/Example/Supporting Files/Assets.xcassets/Contents.json
